requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
customized-konlpy==0.0.64
cycler==0.11.0
django-heroku==0.3.1
fonttools==4.33.3
h11==0.13.0
idna==3.3
joblib==1.1.0
JPype1==1.4.0
kiwisolver==1.4.3
konlpy==0.6.0
lxml==4.9.0
matplotlib==3.5.2
more-itertools==8.13.0
mysqlclient==2.1.0
nltk==3.7
numpy==1.22.4
outcome==1.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
psutil==5.9.1
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
soynlp==0.0.493
threadpoolctl==3.1.0
tika==1.24
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wordcloud==1.8.1
wsproto==1.1.0
